feat(refactor): move trainData to data/ folder
